+++ short-read-mngs/idseq-dag/idseq_dag/steps/generate_annotated_fasta.py
@@ -6,6 +6,7 @@ import idseq_dag.util.m8 as m8
 from idseq_dag.engine.pipeline_step import PipelineCountingStep
 from idseq_dag.util.idseq_dedup_clusters import parse_clusters_file
 from idseq_dag.util.count import READ_COUNTING_MODE, ReadCountingMode
+from idseq_dag.util.parsing import BlastnOutput6Reader
 
 UNMAPPED_HEADER_PREFIX = '>NR::NT::'
 
@@ -69,7 +70,7 @@ class PipelineStepGenerateAnnotatedFasta(PipelineCountingStep):
     def annotate_fasta_with_accessions(merged_input_fasta, nt_m8, nr_m8, output_fasta):
         def get_map(blastn_6_path):
             with open(blastn_6_path) as blastn_6_f:
-                return {row["qseqid"]: row["sseqid"] for row in m8.BlastnOutput6Reader(blastn_6_f, filter_invalid=True)}
+                return {row["qseqid"]: row["sseqid"] for row in BlastnOutput6Reader(blastn_6_f, filter_invalid=True)}
 
         nt_map = get_map(nt_m8)
         nr_map = get_map(nr_m8)
