+++ short-read-mngs/idseq-dag/idseq_dag/util/m8.py
@@ -2,12 +2,9 @@ import json
 import math
 import os
 import random
-import csv
 
-from abc import ABC, abstractstaticmethod
 from collections import Counter
 from collections import defaultdict
-from typing import Any, Callable, Dict, Iterable, List, TextIO, Tuple
 
 import idseq_dag.util.command as command
 import idseq_dag.util.lineage as lineage
@@ -15,6 +12,7 @@ import idseq_dag.util.log as log
 
 from idseq_dag.util.count import READ_COUNTING_MODE, ReadCountingMode, get_read_cluster_size, load_duplicate_cluster_sizes
 from idseq_dag.util.dict import open_file_db_by_extension
+from idseq_dag.util.parsing import BlastnOutput6Reader, BlastnOutput6Writer, HitSummaryReader, HitSummaryWriter
 
 # NT alginments with shorter length are associated with a high rate of false positives.
 # NR doesn't have this problem because Rapsearch2 contains an equivalent filter.
@@ -28,224 +26,6 @@ MAX_EVALUE_THRESHOLD = 1
 # The minimum read count for a valid contig. We ignore contigs below this read count in most downstream analyses.
 # This constant is hardcoded in at least 4 places in idseq-web.  TODO: Make it a DAG parameter.
 MIN_CONTIG_SIZE = 4
-
-Schema = List[Tuple[str, type]]
-
-class _TSVWithSchmaBase(ABC):
-    """
-    The _TSVWithSchemaBase class is a base class for reading and writing TSVs without headers with schemas provided by subclasses
-
-    This class takes a stream for reading or writing as well as a list of partial schemas. These schemas are used to construct a
-    map from the number of fields to the appropriate schema. The first partial schema becomes the first complete schema
-    and each subsequent schema is appended to the previous complete schema to make a complete schema. This guaruntees that
-    there is exactly one schema for each field count. This schema recognition is necessary because previously we were using
-    optional fields to capture a finite set of different variants. This base class just handles constructing the schema map
-    and accessing schemas and fields.
-
-    Args:
-        tsv_stream (TextIO): a text stream to write to/read from
-        *schemas (List[Tuple[str, Callable[[str], Any]]]): partial schemas
-    """
-
-    @abstractstaticmethod
-    def _variants() -> List[Tuple[str, Schema]]:
-        pass
-
-    @classmethod
-    def _build_variant_map(cls) -> Tuple[Dict[str, Schema], Dict[int, str], Dict[str, str]]:
-        schemas = cls._variants()
-        assert schemas, "_TSVWithSchemaReader requires at least one schema"
-        n = 0
-        variants = {"dummy": []}
-        num_fields_to_variant = {0: "dummy"}
-        field_to_first_variant = {}
-        for (variant_name, schema) in schemas:
-            assert schema, "_TSVWithSchemaBase does not support empty schemas"
-            n += len(schema)
-            variants[variant_name] = variants[num_fields_to_variant[n - len(schema)]] + schema
-            num_fields_to_variant[n] = variant_name
-            for field, _ in schema:
-                field_to_first_variant[field] = variant_name
-        del variants["dummy"]
-        del num_fields_to_variant[0]
-        return variants, num_fields_to_variant, field_to_first_variant
-
-    @classmethod
-    def fields(cls, variant_name) -> List[str]:
-        return [k for k, _ in cls._build_variant_map()[0][variant_name]]
-
-    def __init__(self, tsv_stream: TextIO) -> None:
-        self._tsv_stream = tsv_stream
-        self._variants, self._num_fields_to_variant, self._field_to_first_variant = self._build_variant_map()
-        self._schema = None
-
-class _TSVWithSchemaReader(_TSVWithSchmaBase, ABC):
-    def __init__(self, tsv_stream: TextIO, *schemas: List[Tuple[str, Callable[[str], Any]]]) -> None:
-        super().__init__(tsv_stream, *schemas)
-        self._generator = ({
-            key: _type(value) for ((key, _type), value) in zip(self._get_schema(row), row)
-        } for row in csv.reader(self._tsv_stream, delimiter="\t"))
-
-    def __iter__(self):
-        return self
-
-    def __next__(self, *args) -> Dict[str, Any]:
-        return next(self._generator, *args)
-
-    def _get_schema(self, row):
-        if self._schema:
-            return self._schema
-        if len(row) not in self._num_fields_to_variant:
-            raise Exception(f"_TSVWithSchemaReader error. Input: \"{row}\" has {len(row)} fields, no associated schema found in {self._variants}")
-        return self._variants[self._num_fields_to_variant[len(row)]]
-
-class _TSVWithSchemaWriter(_TSVWithSchmaBase, ABC):
-    def __init__(self, tsv_stream: TextIO, *schemas: List[Tuple[str, Callable[[str], Any]]]) -> None:
-        super().__init__(tsv_stream, *schemas)
-        self._writer = csv.writer(tsv_stream, delimiter="\t")
-
-    def _get_schema(self, row):
-        if self._schema:
-            return self._schema
-        if len(row) in self._num_fields_to_variant:
-            return self._variants[self._num_fields_to_variant[len(row)]]
-        first_inclusive_variant = max(
-            (self._field_to_first_variant.get(field, '') for field in row.keys()),
-            key=lambda variant: len(self._variants[variant] if variant else 0)
-        )
-        if first_inclusive_variant:
-            raise Exception(f"_TSVWithSchemaWriter error. Input: \"{row}\" has {len(row)} fields, no associated schema or common fields found in {self._variants}")
-        return self._variants[first_inclusive_variant]
-
-    def _dict_row_to_list(self, row: Dict[str, Any]) -> List[Any]:
-        return [row.get(key) for (key, _) in self._get_schema(row)]
-
-    def write_all(self, rows: Iterable[Dict[str, Any]]) -> None:
-        self._writer.writerows(self._dict_row_to_list(row) for row in rows)
-
-    def write(self, row: Dict[str, Any]) -> None:
-        self._writer.writerow(self._dict_row_to_list(row))
-
-
-# blastn output format 6 as documented in
-# http://www.metagenomics.wiki/tools/blast/blastn-output-format-6
-# it's also the format of our GSNAP and RAPSEARCH2 output
-_BLAST_OUTPUT_SCHEMA = [
-    ("qseqid", str),
-    ("sseqid", str),
-    ("pident", float),
-    ("length", int),
-    ("mismatch", int),
-    ("gapopen", int),
-    ("qstart", int),
-    ("qend", int),
-    ("sstart", int),
-    ("send", int),
-    ("evalue", float),
-    ("bitscore", float),
-]
-
-
-# Additional blastn output columns.
-_BLAST_OUTPUT_NT_SCHEMA = [
-    ("qlen", int),      # query sequence length, helpful for computing qcov
-    ("slen", int),      # subject sequence length, so far unused in IDseq
-]
-
-
-# Re-ranked output of blastn.  One row per query.  Two additional columns.
-_RERANKED_BLAST_OUTPUT_NT_SCHEMA = [
-    ("qcov", float),     # fraction of query covered by the optimal set of HSPs
-    ("hsp_count", int),   # cardihnality of optimal fragment cover;  see BlastCandidate
-]
-
-class _BlastnOutput6Base(ABC):
-    @staticmethod
-    def _variants():
-        return [
-            ("base", _BLAST_OUTPUT_SCHEMA),
-            ("nt", _BLAST_OUTPUT_NT_SCHEMA),
-            ("reranked_nt", _RERANKED_BLAST_OUTPUT_NT_SCHEMA),
-        ]
-
-class BlastnOutput6Reader(_BlastnOutput6Base, _TSVWithSchemaReader):
-    def __init__(self, tsv_stream: TextIO, filter_invalid: bool = False, min_alignment_length: int = 0):
-        if filter_invalid:
-            # The output of rapsearch2 contains comments that start with '#', these should be skipped
-            super().__init__(line for line in tsv_stream if line and line[0] != "#")
-            self._generator = (row for row in self._generator if self.row_is_valid(row, min_alignment_length))
-        else:
-            super().__init__(tsv_stream)
-
-    @staticmethod
-    def row_is_valid(row, min_alignment_length) -> bool:
-        # GSNAP outputs bogus alignments (non-positive length /
-        # impossible percent identity / NaN e-value) sometimes,
-        # and usually they are not the only assignment, so rather than
-        # killing the job, we just skip them. If we don't filter these
-        # killing the job, we just skip them. If we don't filter these
-        # out here, they will override the good data when computing min(
-        # evalue), pollute averages computed in the json, and cause the
-        # webapp loader to crash as the Rails JSON parser cannot handle
-        # NaNs. Test if e_value != e_value to test if e_value is NaN
-        # because NaN != NaN.
-        # *** E-value Filter ***
-        # Alignments with e-value > 1 are low-quality and associated with false-positives in
-        # all alignments steps (NT and NR). When the e-value is greater than 1, ignore the
-        # alignment
-        ###
-        return all([
-            row["length"] >= min_alignment_length,
-            -0.25 < row["pident"] < 100.25,
-            row["evalue"] == row["evalue"],
-            row["evalue"] <= MAX_EVALUE_THRESHOLD,
-        ])
-
-class BlastnOutput6Writer(_BlastnOutput6Base, _TSVWithSchemaWriter):
-    pass
-
-
-_HIT_SUMMARY_SCHEMA = [
-    ("read_id", str),
-    ("level", int),
-    ("taxid", int),
-    ("accession_id", str),
-    ("species_taxid", int),
-    ("genus_taxid", int),
-    ("family_taxid", int),
-]
-
-_HIT_SUMMARY_CONTIG_SCHEMA = [
-    ("contig_id", str),
-    ("contig_accession_id", str),
-    ("contig_species_taxid", int),
-    ("contig_genus_taxid", int),
-    ("contig_family_taxid", int),
-]
-
-_HIT_SUMMARY_ASSEMBLY_SOURCE_SCHEMA = [
-    ("from_assembly", str),
-]
-
-_HIT_SUMMARY_MERGED_SCHEMA = [
-    ("source_count_type", str),
-]
-
-class _HitSummaryBase(ABC):
-    @staticmethod
-    def _schemas():
-        return [
-            ("base", _HIT_SUMMARY_SCHEMA),
-            ("contig", _HIT_SUMMARY_CONTIG_SCHEMA),
-            ("assembly_source", _HIT_SUMMARY_ASSEMBLY_SOURCE_SCHEMA),
-            ("merged", _HIT_SUMMARY_MERGED_SCHEMA),
-        ]
-
-class HitSummaryReader(_HitSummaryBase, _TSVWithSchemaReader):
-    pass
-
-class HitSummaryWriter(_HitSummaryBase, _TSVWithSchemaWriter):
-    pass
 
 
 def summarize_hits(hit_summary_file_path: str, min_reads_per_genus=0):
